requirements.txt
matplotlib>=3.0.0
numpy>=1.19.0
pandas>=1.0.0
seaborn>=0.11.0
nltk>=3.6.0
scipy>=1.6.0
tqdm>=4.60.0
openai>=1.0.0
google-generativeai>=0.1.0
typing-extensions>=4.0.0